+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/PriceModifierRepository.java
@@ -8,11 +8,11 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface PriceModifierRepository extends JpaRepository<PriceModifier, Long> {
+public interface PriceModifierRepository extends JpaRepository<PriceModifier, Integer> {
 
     @Query("FROM PriceModifier pm WHERE pm.owner.id =:ownerId AND pm.startDate <:time AND pm.endDate >:time")
     List<PriceModifier> getPriceModifierByProductAndDates(
-            @Param(value = "ownerId") long ownerId,
+            @Param(value = "ownerId") int ownerId,
             @Param(value = "time")LocalDateTime time);
 
     PriceModifier findByName(String name);
