+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/ReceiptRecordCreatedRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 import java.time.LocalDateTime;
 import java.util.List;
 
-public interface ReceiptRecordCreatedRepository extends JpaRepository<ReceiptRecordCreated, Long> {
+public interface ReceiptRecordCreatedRepository extends JpaRepository<ReceiptRecordCreated, Integer> {
 
     @Query("FROM ReceiptRecordCreated rrc WHERE rrc.created >:created AND rrc.owner.name=:name ORDER BY rrc.created DESC")
     List<ReceiptRecordCreated> findRecentByTimestamp(@Param(value = "name") String name, @Param(value = "created") LocalDateTime created);
