+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/RestaurantRepository.java
@@ -3,5 +3,5 @@ package com.inspirationlogical.receipt.corelib.repository;
 import com.inspirationlogical.receipt.corelib.model.entity.Restaurant;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RestaurantRepository extends JpaRepository<Restaurant, Long> {
+public interface RestaurantRepository extends JpaRepository<Restaurant, Integer> {
 }
