+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/VATRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface VATRepository extends JpaRepository<VAT, Long> {
+public interface VATRepository extends JpaRepository<VAT, Integer> {
 
     @Query("FROM VAT v WHERE v.name = :name AND v.status = 'VALID'")
     VAT getVatByName(@Param(value = "name") VATName name);
