+++ app/corelib/src/main/java/com/inspirationlogical/receipt/corelib/repository/VATSerieRepository.java
@@ -4,7 +4,7 @@ import com.inspirationlogical.receipt.corelib.model.entity.VATSerie;
 import com.inspirationlogical.receipt.corelib.model.enums.VATStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VATSerieRepository extends JpaRepository<VATSerie, Long> {
+public interface VATSerieRepository extends JpaRepository<VATSerie, Integer> {
 
     VATSerie findFirstByStatus(VATStatus status);
 }
